fixed main typos

+++ main.cpp
@@ -32,7 +32,7 @@ int main()
     for (int i = 0; i < 1000; i++)
     {
         usleep(100);
-        int lux = ltr559.getLux();
+        float lux = ltr559.getLux();
         std::cout << "Luminosité : " << lux << " lux" << std::endl;
     }
     // initializing BME280
